webhook tests: move test suite to own file
In preparation for adding more API resources with their own validation,
move the test suite to their own files.
